+++ src/tests/keypair.test.ts
@@ -12,7 +12,7 @@ import { exec as execNoPromises } from "node:child_process";
 import { promisify } from "node:util";
 import { writeFile, unlink as deleteFile } from "node:fs/promises";
 import dotenv from "dotenv";
-import { createSignerFromKeyPair, KeyPairSigner, createKeyPairSignerFromBytes } from "@solana/web3.js";
+import { createSignerFromKeyPair, KeyPairSigner, createKeyPairSignerFromBytes } from "@solana/kit";
 import { log } from "../lib/serializer";
 
 const exec = promisify(execNoPromises);