+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `getPDAAndBump()` - calculates a Program Derived Address (PDA) and its bump seed from a program address and seeds, automatically encoding different seed types (strings, addresses, and bigints).
 - Add TSDoc comments for all functions.
 - Fix bug where `mintTokens()` was minting to the mint authority rather than the destination.
+- Solana `@solana/kit` has been renamed to `@solana/kit`
 
 ## Kite version 1.01